Tp_Laboratorio_1

+++ TP N° 1/Funciones.h
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 
-/** \brief float Hace la suma de los numeros ingresados
+/** \brief float Hace la suma de dos numeros ingresados
  *
  * \param float Uno de los valores al cual se le aplica la operacion
  * \param float Otro de los valores
@@ -11,7 +11,7 @@
  */
 float sumarNumeros(float, float);
 
-/** \brief float Hace la resta de los numeros ingresados
+/** \brief float Hace la resta de dos numeros ingresados
  *
  * \param float Uno de los numeros que participa en la operacion
  * \param float Otro de los numeros que tambien participa
@@ -20,7 +20,7 @@ float sumarNumeros(float, float);
  */
 float restarNumeros(float, float);
 
-/** \brief float Realiza la multiplicacion de los numeros ingresados
+/** \brief float Realiza la multiplicacion de dos numeros ingresados
  *
  * \param float Uno de los numeros que participa en la operacion
  * \param float Otro de los numeros que tambien participa en la operacion
@@ -29,7 +29,7 @@ float restarNumeros(float, float);
  */
 float multiplicarNumeros(float, float);
 
-/** \brief float Realiza la division de dos numeros enteros
+/** \brief float Realiza la division de dos numeros ingresados
  *
  * \param float Uno de los numeros que participa en la operacion
  * \param float Otro de los numeros que tambien participa en la operacion
@@ -44,7 +44,7 @@ float dividirNumeros(float, float);
  * \return int El resultado del factorial del numero
  *
  */
-int factorizarNumeroA(int );
+long int factorizarNumeroA(int );
 
 /** \brief int Calcula el factorial de un numero entero
  *
@@ -52,12 +52,12 @@ int factorizarNumeroA(int );
  * \return int El resultado del factorial del numero
  *
  */
-int factorizarNumeroB(int );
+long int factorizarNumeroB(int );
 
 /** \brief int Contiene el menu con las opciones a elejir
  *
- * \param
- * \param
+ *
+ *
  * \return la opcion que se elijio
  *
  */
